done with part4 exercises all of it

+++ controllers/blogrouter.js
@@ -1,29 +1,69 @@
 const blogRouter = require('express').Router()
+const blog = require('../models/blog')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogRouter.get('/',  async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
   response.status(200).json(blogs)
 
 })
 
-blogRouter.post('/', (request, response, next) => {
-    const blog = new Blog(request.body)
-  
-    blog
-      .save()
-      .then(result => {
-        response.status(201).json(result)
-      })
-      .catch(error => next(error))
+
+
+blogRouter.post('/', async (request, response) => {
+  const {title, author, url, likes} = request.body
+  const user = request.user
+  if (!user) {
+    return response.status(401).json({error: 'token invalid'})
+  }
+
+  // 
+
+  const blog = new Blog({
+    title,
+    author,
+    url,
+    likes,
+    user: user._id
   })
 
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.status(201).json(savedBlog)
+
+})
+
 blogRouter.delete('/:id', async (request, response) => {
-  const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
-  if (!deletedBlog) {
+  const user = request.user
+  
+
+  const blog = await Blog.findById(request.params.id)
+  
+  if (!blog) {
     return response.status(404).json({error: 'Blog not found'})
   }
-  response.status(204).end()
+  if (!blog.user) {
+    return response.status(401).send({error: 'Unauthorized to delete this blog'})
+  }
+
+  
+
+
+
+  if (blog.user.toString() === user._id.toString()){
+    
+    const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
+    if (!deletedBlog) {
+      return response.status(404).json({error: 'error deleting blog'})
+    }
+    // 
+    return response.status(204).end()
+  } else {
+    return response.status(401).json({error: 'Unauthorized to delete this blog'})
+  }
 })
 
 
@@ -39,7 +79,7 @@ blogRouter.put('/:id', async (request, response) => {
 
 
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new: true, runValidators: true})
-  console.log(updatedBlog)
+  
 
   if (!updatedBlog){
     // const error = new Error('Blog not found')
